using Modals to display the charts in zoom mode
